rename resources package to models *resources* is misleading - more accurately, these classes describe pojo models
