+++ README.md
@@ -29,8 +29,8 @@ import com.github.jasminb.jsonapi.JSONAPIDocument;
 import com.patreon.PatreonAPI;
 import com.patreon.PatreonOAuth;
 import com.patreon.PatreonOAuth;
-import com.patreon.resources.User;
-import com.patreon.resources.Pledge;
+import com.patreon.models.User;
+import com.patreon.models.Pledge;
 
     ...
