+++ src/main/java/com/patreon/PatreonAPI.java
@@ -3,12 +3,12 @@ package com.patreon;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.PropertyNamingStrategy;
 import com.github.jasminb.jsonapi.*;
-import com.patreon.resources.Campaign;
-import com.patreon.resources.Pledge;
-import com.patreon.resources.RequestUtil;
-import com.patreon.resources.User;
-import com.patreon.resources.shared.BaseResource;
-import com.patreon.resources.shared.Field;
+import com.patreon.models.Campaign;
+import com.patreon.models.Pledge;
+import com.patreon.models.RequestUtil;
+import com.patreon.models.User;
+import com.patreon.models.shared.BaseResource;
+import com.patreon.models.shared.Field;
 import org.apache.http.NameValuePair;
 import org.apache.http.client.utils.URIBuilder;
 import org.apache.http.client.utils.URLEncodedUtils;
