+++ src/main/java/com/patreon/models/User.java
@@ -1,12 +1,12 @@
-package com.patreon.resources;
+package com.patreon.models;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.github.jasminb.jsonapi.annotations.Relationship;
 import com.github.jasminb.jsonapi.annotations.Type;
-import com.patreon.resources.shared.BaseResource;
-import com.patreon.resources.shared.Field;
-import com.patreon.resources.shared.SocialConnections;
+import com.patreon.models.shared.BaseResource;
+import com.patreon.models.shared.Field;
+import com.patreon.models.shared.SocialConnections;
 
 import java.util.Arrays;
 import java.util.Collection;
