+++ src/main/java/com/patreon/models/shared/SocialConnections.java
@@ -1,4 +1,4 @@
-package com.patreon.resources.shared;
+package com.patreon.models.shared;
 
 
 import java.util.List;
