+++ src/test/java/com/patreon/PatreonAPITest.java
@@ -1,17 +1,15 @@
 package com.patreon;
 
 import com.github.jasminb.jsonapi.JSONAPIDocument;
-import com.patreon.resources.Campaign;
-import com.patreon.resources.Pledge;
-import com.patreon.resources.RequestUtil;
-import com.patreon.resources.User;
+import com.patreon.models.Campaign;
+import com.patreon.models.Pledge;
+import com.patreon.models.RequestUtil;
+import com.patreon.models.User;
 import junit.framework.TestCase;
 import org.apache.http.NameValuePair;
 import org.apache.http.client.utils.URLEncodedUtils;
 import org.mockito.ArgumentCaptor;
-import org.mockito.Matchers;
 import org.mockito.Mockito;
-import org.powermock.api.mockito.PowerMockito;
 
 import java.nio.charset.Charset;
 import java.util.Arrays;
